refactor(workflow-executor): co-locate the MCP plane under src/mcp
Move the MCP types, errors and router out of the shared root files into the
src/mcp module (types.ts, errors.ts, mcp-router.ts) alongside the service and
resolver. The MCP errors still extend the root domain categories so toHttpError
maps them unchanged. Root errors.ts and types/ no longer carry MCP concerns.
