+++ packages/workflow-executor/src/mcp/mcp-permission-resolver.ts
@@ -1,4 +1,4 @@
-import type { CanExecuteOperation, CanExecuteVerdict } from '../types/mcp';
+import type { CanExecuteOperation, CanExecuteVerdict } from './types';
 
 export interface McpPermissionResolver {
   resolve(userId: number, operation: CanExecuteOperation): Promise<CanExecuteVerdict>;
